--- projects/server/config/config.json
@@ -1,24 +1,0 @@
-{
-  "development": {
-    "username": "root",
-    "password": "ddd1234",
-    "database": "db_warehouse",
-    "host": "127.0.0.1",
-    "dialect": "mysql",
-    "timezone": "+07:00"
-  },
-  "test": {
-    "username": "root",
-    "password": null,
-    "database": "database_test",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  },
-  "production": {
-    "username": "root",
-    "password": null,
-    "database": "database_production",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  }
-}
