+++ projects/server/connection/conn.js
@@ -4,13 +4,13 @@ const mysql = require('mysql2')
 const db = mysql.createConnection({
     host: 'localhost',
     user: 'root',
-    password: 'ddd1234',
+    password: 'ot3ra1han12',
     database: 'db_warehouse',
-    port: 3306
+    port: 8000
 })
 
 db.connect((err) => {
-    if(err) return console.log('Error ' + err.message)
+    if (err) return console.log('Error ' + err.message)
     else console.log('MySQL Connected')
 })
