+++ projects/server/sequelize/models/index.js
@@ -1,0 +1,39 @@
+'use strict';
+
+const fs = require('fs');
+const path = require('path');
+const Sequelize = require('sequelize');
+const process = require('process');
+const basename = path.basename(__filename);
+const env = process.env.NODE_ENV || 'development';
+const config = require(__dirname + '/../config/config.js').db[env];
+const db = {};
+const JSONConfig = JSON.parse(JSON.stringify(config))
+
+let sequelize;
+if (JSONConfig.use_env_variable) {
+    sequelize = new Sequelize(process.env[JSONConfig.use_env_variable], JSONConfig);
+} else {
+    sequelize = new Sequelize(JSONConfig.database, JSONConfig.username, JSONConfig.password, JSONConfig);
+}
+
+fs
+    .readdirSync(__dirname)
+    .filter(file => {
+        return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js');
+    })
+    .forEach(file => {
+        const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
+        db[model.name] = model;
+    });
+
+Object.keys(db).forEach(modelName => {
+    if (db[modelName].associate) {
+        db[modelName].associate(db);
+    }
+});
+
+db.sequelize = sequelize;
+db.Sequelize = Sequelize;
+
+module.exports = db;
