+++ projects/server/sequelize/models/product.js
